Move struct config to its own file
